型を考慮したプログラムに修正

+++ RouteGenerator.cpp
@@ -10,7 +10,7 @@ Array<Command> RouteGenerator::getCommands() const
 	Array<Command> commands;
 	for (const auto& p : m_route)
 	{
-		commands.emplace_back('P', p.x * 10, 2000, p.y * 10, 2000);
+		commands.emplace_back('P', short(p.x * 10), short(2000), short(p.y * 10), short(2000));
 	}
 
 	return commands;
